+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,19 @@
 
 ## Next Release
 
+## 0.2.5-beta.17
+
+- \[contracts\] Use pending transaction count for nonce
+
+## 0.2.5-beta.16
+
+- \[contracts\] Bump gas limit
+- \[test-runner\] Working load tests
+
+## 0.2.5-beta.14
+
+- \[contracts\] Add additional error message for nonce handling
+
 ## 0.2.5-beta.13
 
 - \[engine\] hotfix for withdraw retry schema